+++ examples/futures/trade/getCurrentPositionRisk.js
@@ -1,0 +1,15 @@
+'use strict'
+
+const BitmartFuturesAPI = require('../../../src/bitmartFuturesAPI')
+
+const yourApiKey = 'your api key'
+const bitmartFuturesAPI = new BitmartFuturesAPI({
+  apiKey: yourApiKey
+})
+
+
+bitmartFuturesAPI.getCurrentPositionRisk({
+    symbol: 'ETHUSDT'
+})
+  .then(response => bitmartFuturesAPI.logger.log(response.data))
+  .catch(error => bitmartFuturesAPI.logger.log(error))