+++ examples/spot/account/withdraw.js
@@ -12,9 +12,22 @@ const bitmartSpotAPI = new BitmartSpotAPI({
   apiMemo: yourApiMemo
 })
 
+// Withdraw to the blockchain
+bitmartSpotAPI.withdraw('USDT-TRC20', '0.001', {
+    address:'0x1EE6FA5A3803608fc22a1f3F76',
+    address_memo:'',
+    destination:'To Digital Address'
+})
+  .then(response => bitmartSpotAPI.logger.log(response.data))
+  .catch(error => bitmartSpotAPI.logger.log(error))
 
-bitmartSpotAPI.withdraw('USDT-TRC20', '0.001', 'To Digital Address', '0x1EE6FA5A3803608fc22a1f3F76', {
-    address_memo:''
+
+
+//  Withdraw to BitMart account
+bitmartSpotAPI.withdraw('USDT-TRC20', '0.001', {
+  type:1,
+  value:'123424',
+  areaCode:''
 })
   .then(response => bitmartSpotAPI.logger.log(response.data))
   .catch(error => bitmartSpotAPI.logger.log(error))