+++ examples/spot/trade/cancelAllOrde.js
@@ -12,24 +12,12 @@ const bitmartSpotAPI = new BitmartSpotAPI({
     apiMemo: yourApiMemo,
 })
 
-const params = [{
-    "size": "0.1",
-    "price": "8800",
-    "side": "buy",
-    "type": "limit",
-    "clientOrderId":"xxdfs23e232"
-}, {
-    "size": "0.1",
-    "price": "8800",
-    "side": "sell",
-    "type": "limit",
-    "clientOrderId":"xxdfs23e233"
-}];
 
-bitmartSpotAPI.v4NewBatchSpotOrder('BTC_USDT', params, {
-    recvWindow: 5000
-})
-.then(response => bitmartSpotAPI.logger.log(response.data))
+
+bitmartSpotAPI.v4CancelAllOrder({
+    symbol: 'BTC_USDT',
+    side: 'buy'
+}).then(response => bitmartSpotAPI.logger.log(response.data))
     .catch(error => {
         if (error.response) {
             bitmartSpotAPI.logger.log(error.response.data);