+++ examples/spot/websocket/subscribePublicChannel.js
@@ -6,19 +6,21 @@ const BitmartSpotWebsocket = require('../../../src/bitmartSpotWebsocket')
 const callbacks = {
   open: (client) => {
     // 【Public】Ticker Channel
-    client.send('{"op": "subscribe", "args": ["spot/ticker:BTC_USDT"]}')
+    // client.send('{"op": "subscribe", "args": ["spot/ticker:BTC_USDT"]}')
 
     // 【Public】KLine Channel
-    client.send('{"op": "subscribe", "args": ["spot/kline1m:BTC_USDT"]}')
+    // client.send('{"op": "subscribe", "args": ["spot/kline1m:BTC_USDT"]}')
 
     // 【Public】Depth Channel
-    client.send('{"op": "subscribe", "args": ["spot/depth5:BTC_USDT"]}')
+    // client.send('{"op": "subscribe", "args": ["spot/depth5:BTC_USDT"]}')
 
     // 【Public】Trade Channel
-    client.send('{"op": "subscribe", "args": ["spot/trade:BTC_USDT"]}')
+    // client.send('{"op": "subscribe", "args": ["spot/trade:BTC_USDT"]}')
 
   },
   close: () => console.info('.........Disconnected with Websocket server'),
+  pong: () => console.info('recv:pong from server'),
+  ping: () => console.info('recv:ping from server'),
   message: data => console.info('recv:' + data)
 }
 
@@ -27,5 +29,5 @@ const bitmartSpotWebsocket = new BitmartSpotWebsocket(
     callbacks: callbacks 
 })
 
-// disconnect after 20 seconds
-setTimeout(() => bitmartSpotWebsocket.disconnect(), 20000)
+// If it is a test, you can turn on active shutdown, disconnect after 20 seconds
+// setTimeout(() => bitmartSpotWebsocket.disconnect(), 20000)