+++ CHANGELOG.md
@@ -3,6 +3,23 @@ Changelog
 
 ### v2.0.0 Release
 #### New Features
+- New
+  - `/spot/v4/batch_orders`
+  - `/spot/v4/cancel_orders`
+  - `/spot/v4/cancel_all`
+  - `/contract/private/current-plan-order`
+  - `/contract/private/position-risk`
+- Updated 
+  - `/account/v1/withdraw/apply`
+- Removed
+  - `/spot/v2/ticker`
+  - `/spot/v1/ticker_detail`
+  - `/spot/v1/steps`
+  - `/spot/v1/symbols/kline`
+  - `/spot/v1/symbols/book`
+  - `/spot/v1/symbols/trades`
+  - `/spot/v2/batch_orders`
+  - `/spot/v1/cancel_orders`
 #### Improvements
 - Support custom request headers
 #### Bug Fixes